🦶 base screen from AI

+++ firmware/src/core/IPower.h
@@ -9,6 +9,6 @@ class IPower
 public:
   virtual ~IPower() = default;
   virtual void begin() = 0;
-  virtual int getBatteryPercentage() = 0;
+  virtual uint8_t getBatteryPercentage() = 0;
   virtual void powerOff() = 0;
 };
